requirements.txt
streamlit
google-genai
firebase-admin
pillow
google-cloud-speech
google-cloud-translate==2.0.2
google-cloud-storage
streamlit-audiorec
pydub
ffmpeg

google-cloud-vision

google-api-python-client
google-auth-httplib2
google-auth-oauthlib